Move SignalRemovalArgs to RemoveSignal.h
