+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -138,42 +138,6 @@ namespace OpenLoco::GameCommands
         return doCommand(T::command, regs);
     }
 
-    struct SignalRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeSignal;
-
-        SignalRemovalArgs() = default;
-        explicit SignalRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-            , rotation(regs.bh & 0x3)
-            , trackId(regs.dl & 0x3F)
-            , index(regs.dh & 0xF)
-            , type(regs.bp & 0xF)
-            , flags(regs.edi >> 16)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-        uint8_t trackId;
-        uint8_t index;
-        uint8_t type;
-        uint16_t flags;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.edi = pos.z | (flags << 16);
-            regs.bh = rotation;
-            regs.dl = trackId;
-            regs.dh = index;
-            regs.bp = type;
-            return regs;
-        }
-    };
-
     struct TrainStationRemovalArgs
     {
         static constexpr auto command = GameCommand::removeTrainStation;
