+++ src/OpenLoco/src/Windows/Construction/SignalTab.cpp
@@ -3,6 +3,7 @@
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Track/CreateSignal.h"
+#include "GameCommands/Track/RemoveSignal.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
 #include "Localisation/FormatArguments.hpp"
